shops - all the logic moved to shops service

+++ backend/src/shops/shops.service.ts
@@ -1,10 +1,11 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Shop, ShopRole } from './shop.entity';
 import * as bcrypt from 'bcrypt';
 import { CreateShopDto } from './dto/create-shop.dto';
 import { UpdateShopDto } from './dto/update-shop.dto';
+import { JwtShop } from '../auth/jwt-shop.type';
 
 @Injectable()
 export class ShopsService {
@@ -33,7 +34,15 @@ export class ShopsService {
     return result as Shop;
   }
 
-  async updateShop(id: string, dto: UpdateShopDto): Promise<Shop> {
+  async updateShop(user: JwtShop, id: string, dto: UpdateShopDto): Promise<Shop> {
+    if (user.role !== ShopRole.CEO && user.shopId !== id) {
+      throw new ForbiddenException('You are not allowed to update this shop');
+    }
+    
+    if (dto.role && user.role !== ShopRole.CEO) {
+      throw new ForbiddenException('Only CEO can change shop roles');
+    }
+
     const shop = await this.shopRepository.findOne({ where: { id } });
     if (!shop) {
       throw new NotFoundException(`Shop with id ${id} not found`);
@@ -54,7 +63,10 @@ export class ShopsService {
     return result as Shop;
   }
 
-  async deleteShop(id: string): Promise<void> {
+  async deleteShop(user: JwtShop, id: string): Promise<void> {
+    if (user.role !== ShopRole.CEO && user.shopId !== id) {
+      throw new ForbiddenException('You are not allowed to delete this shop');
+    }
     const result = await this.shopRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Shop with id ${id} not found`);
@@ -65,7 +77,10 @@ export class ShopsService {
     return this.shopRepository.find({select: ['id', 'name', 'email', 'role', 'createdAt', 'updatedAt']});
   }
 
-  async findOne(id: string): Promise<Shop> {
+  async findById(user: JwtShop, id: string): Promise<Shop> {
+    if (user.role !== ShopRole.CEO && user.shopId !== id) {
+      throw new ForbiddenException('You are not allowed to fetch another shop info');
+    }
     const shop = await this.shopRepository.findOne({
          where: { id },
          select: ['id', 'name', 'email', 'role', 'createdAt', 'updatedAt'], 
@@ -84,7 +99,10 @@ export class ShopsService {
     return shop;
   }
 
-  async findByName(name: string): Promise<Shop> {
+  async findByName(user: JwtShop, name: string): Promise<Shop> {
+    if (user.role !== ShopRole.CEO && user.name !== name) {
+      throw new ForbiddenException('You are not allowed to fetch another shop info');
+    }
     const shop = await this.shopRepository.findOne({
       where: { name },
       select: ['id', 'name', 'email', 'role', 'createdAt', 'updatedAt'],
